refactor(787): finish the composition root -- move controller construction into bootstrap.ts (phase 10b-g)
Moves host/panelContext/ceremonies/diplomacyActions/panelActions/
selectionController/turnFlow/playerActions/mapInteraction/hud/campaignEntry/
gameSession/presentationContext/panelRegistry/router construction out of
main.ts's module scope into a new createAppComposition factory in
src/app/bootstrap.ts. main.ts keeps only the true externally-supplied
primitives, the notifier let (now threaded via getNotifier/setNotifier),
and the Phase-13-scoped functions (foundCityAction, beginPlayerCityAssault,
executeAttack, executeMinorCivConquest, executeUpgrade,
maybeShowPendingHoardChoice, showNotification), which now reference the
returned AppComposition instead of bare module-scope bindings.

main.ts: 895 -> 471 lines. bootstrap.ts: 48 -> 559 lines.

Every lazy-wrapper vs. direct-reference decision from the original
construction order is preserved unchanged (verbatim-move technique) --
only the notifier closure became an explicit getNotifier()/setNotifier()
pair since it now crosses a file boundary.

Fixes 3 tests/main.integration.test.ts source-grep assertions that
hardcoded main.ts as the construction site for
createSelectionController(/createMapInteractionController( and the
selectionController-in-executeMinorCivConquest reference.
